Leave extent undefined at creation (otherwise property from C++ are overwritten by QMML defaults)
Fix gravity constant

Weight is distributed over the extent(N/mm)

Force is acting all over the beam in all the three dimensions: Frame3DD
considers beams as 1-dimentional entities, hence the begin-end of
trapezoidal forces on the 3 axes is used to define the components of the
force in the 3 dimensions. If the force magnitude is 0 in one direction,
then it's like the begin-end for that direction was zero.